### 2018-04-21 #689 @whedon
[PRE REVIEW]: GeoStats.jl - High-performance geostatistics in Julia
**Submitting author:** @juliohm (<a href="http://orcid.org/0000-0003-2789-297X">Júlio Hoffimann</a>)
**Repository:** <a href="https://github.com/juliohm/GeoStats.jl" target ="_blank">https://github.com/juliohm/GeoStats.jl</a>
**Version:** v0.6.0
