### 2018-04-21 #689 @juliohm
Thank you @arfon for the follow up. I noticed that the figure in the compiled PDF is placed differently than in my local version. Is there a template that I can use to reproduce exactly the document compiled by the bot?
