### 2018-04-21 bfeaef8
Reduce image size in paper.md
Former-commit-id: 849beff27e1d3917b2bb19fc150ff9ceb0d6ab75
