### 2018-04-21 49b0f8d
Fix English issue in paper.md
Former-commit-id: 4f952b9f64eec4fa73fd5b5b40603c3419c4416c
