### 2018-04-21 #689 @juliohm
I've fixed the figure placement issue, and other small English issue. Looking forward to start the review process.

Best,
