### 2018-04-22 346bdee
Update paper.md
Former-commit-id: 92440d2cb1dec9064e8322899bd5a28cdf7a60e7
