### 2018-04-22 #689 @lheagy
👋 Hi @nuest, would you be willing to review this submission for JOSS?
