### 2018-04-22 #689 @juliohm
Hi @lheagy, thank you for the prompt reply and for pinging @nuest. Is there a list of reviewers from which to select similar to the list of editors on the JOSS website?
