### 2018-04-22 #689 @juliohm
:wave: @dhhagan @ziotom78 @wkearn @rowanc1 would any of you be available to review the submission? I've checked your GitHub profiles and websites, and concluded that you all have experience with spatial data, geosciences & statistics. Looking forward to your feedback.
