### 2018-04-22 f1b3794
Update paper.md
