### 2018-04-23 juliohm/GeoStats.jl#20 @ziotom78
Error in example AnisotropicModels
In the example `AnisotropicModels` (accessible through `GeoStats.examples()`), the following example does not work:
```julia
using Plots; gr()

dim, nobs = 2, 50

X = 100*rand(dim, nobs)
z = rand(nobs)

scatter(X[1,:], X[2,:], c=z, label="data")
```

The error is the following:
```
MethodError: no method matching plot_color(::Float64)
Closest candidates are:
  plot_color(::Any, ::Void) at C:\Users\tomasi\.julia\v0.6\PlotUtils\src\colors.jl:15
  plot_color(::Any, ::Number) at C:\Users\tomasi\.julia\v0.6\PlotUtils\src\colors.jl:18
  plot_color(::Symbol) at C:\Users\tomasi\.julia\v0.6\PlotUtils\src\colors.jl:6
```

Removing the keyword `c=z` makes the error disappear, although the plot might look different from what was intended:
```julia
scatter(X[1,:], X[2,:], label="data")
```

Version information:
```
julia> versioninfo()
Julia Version 0.6.2
Commit d386e40c17* (2017-12-13 18:08 UTC)
Platform Info:
  OS: Windows (x86_64-w64-mingw32)
  CPU: Intel(R) Core(TM) i7-4702MQ CPU @ 2.20GHz
  WORD_SIZE: 64
  BLAS: libopenblas (USE64BITINT DYNAMIC_ARCH NO_AFFINITY Haswell)
  LAPACK: libopenblas64_
  LIBM: libopenlibm
  LLVM: libLLVM-3.9.1 (ORCJIT, haswell)

julia> Pkg.status("GeoStats")
 - GeoStats                      0.6.0              a18650da (dirty)
```
