### 2018-04-23 juliohm/GeoStats.jl#21 @ziotom78
Misleading plot in documentation
The plot shown in the [Quick Example](http://juliohm.github.io/GeoStats.jl/stable/#Quick-example-1) section of the documentation is misleading: if the user runs the code, it will not produce the same plot, as no Gaussian variogram is used. This should be explicitly stated in the documentation, or the plot changed.
