### 2018-04-23 #692 @ziotom78
Really a good and interesting package! I am going through the documentation and the examples, at the moment I have opened two tickets ([#20](https://github.com/juliohm/GeoStats.jl/issues/20) and [#21](https://github.com/juliohm/GeoStats.jl/issues/21)).

In the README, you might consider putting a link to your `CONTRIBUTING.md` file, as it contains valuable information that might be missed.
