### 2018-04-23 juliohm/GeoStats.jl#20 @juliohm
Thank you @ziotom78 for reporting the issue. :+1:  This one is probably due to some update in the Plots.jl package. I will open an issue there to see what they say.
