### 2018-04-23 juliohm/GeoStats.jl#20 @juliohm
Issue opened here: https://github.com/JuliaPlots/Plots.jl/issues/1493
