### 2018-04-23 juliohm/GeoStats.jl#21 @juliohm
Good catch, I will update the plot and the code accordingly :+1:
