### 2018-04-23 #692 @juliohm
Thank you @ziotom78 for your time and for going through the documentation, very helpful. I will add a link to the CONTRIBUTING.md file in the README.md
