### 2018-04-23 8169c47
Update example in README.md
