+++ README.md
@@ -149,7 +149,7 @@ grid = RegularGrid{Float64}(100, 100)
 # define estimation problem for any data column(s) (e.g. :precipitation)
 problem = EstimationProblem(geodata, grid, :precipitation)
 
-# choose a solver from the list
+# choose a solver from the list of solvers
 solver = Kriging(:precipitation => @NT(variogram=GaussianVariogram(range=35.)))
 
 # solve the problem
