### 2018-04-23 86282d0
Update quick example in the docs
Former-commit-id: 7727f5eacf3ef435a57f8eec194ae9dc938e20b9
