### 2018-04-23 a2e4d55
Add link to GeoStatsDevTools.jl in docs main page
