+++ docs/src/index.md
@@ -29,7 +29,8 @@ globe and to standardize this incredible science.
 If you would like to help support the project, please
 [star the repository on GitHub](https://github.com/juliohm/GeoStats.jl)
 and share it with your colleagues. If you are a developer,
-please check [GeoStatsBase.jl](https://github.com/juliohm/GeoStatsBase.jl).
+please check [GeoStatsBase.jl](https://github.com/juliohm/GeoStatsBase.jl)
+and [GeoStatsDevTools.jl](https://github.com/juliohm/GeoStatsDevTools.jl).
 
 ## Installation
 
