### 2018-04-23 8f0fae6
Update example in paper.md
