### 2018-04-23 #692 @juliohm
Fixed the example everywhere, the documentation, the README.md and the paper.md. I've also added the link to CONTRIBUTING.md as suggested to the README.md
