### 2018-04-23 juliohm/GeoStats.jl#20 @juliohm
Updated the notebook with the new syntax, replacing `c=z` with `marker_z=z`.
