### 2018-04-23 juliohm/GeoStats.jl#21 @juliohm
I've updated the plot everywhere, including docs, README.md and the paper.md file.

Thanks again for catching it.
