### 2018-04-24 #692 @ziotom78
Great, @juliohm! The only last thing I think might be improved is the section "Comparison methods" in the paper. It looks a bit sketchy, and the reader might not understand that the library actually implements both comparison methods. You might simply quote the relevant functions and types (`compare`, `VisualComparison`, `CrossValidation`) in the text.
