### 2018-04-24 d41022c
Update paper.md
