### 2018-04-24 #692 @juliohm
Thank you @ziotom78 , I've added a sentence with the `compare` command to clarify the section. It was indeed really sketchy, thank you for your feedback.
