### 2018-04-24 #692 @ziotom78
Excellent! From my point of view the paper as it is now is fine, but let's wait for @wkearn's review.
