### 2018-04-24 #692 @juliohm
Thank you @wkearn! I am glad the documentation and examples are clear. This was a place where I did put a lot of effort. One of my goals is to really demystify geostatistics to the scientific community in general.

I will expand the comparison section of the paper a little more with a figure, and then let you know. We could then wrap up the review process. :+1:
