### 2018-04-24 8b5219a
Update paper.md
Former-commit-id: 73db423721ede30a66b52ef34ba7ec76964198f2
