### 2018-04-24 #692 @juliohm
@ziotom78 @wkearn could you please give your final word on the latest version of the paper above? Please let me know if anything is out of place.
