### 2018-04-25 #692 @lheagy
Thanks @wkearn and @ziotom78 for a speedy and thorough review!

@juliohm: we can proceed to accept the paper - congratulations! Could you please archive your software on [Zenodo](https://zenodo.org/) or similar and post the doi here?
