### 2018-04-25 37315ab
Update version info in paper.md
