### 2018-04-25 64013b4
Add DOI badge to docs
