+++ docs/src/index.md
@@ -8,6 +8,8 @@
 [![License File](https://img.shields.io/badge/license-ISC-blue.svg)](https://github.com/juliohm/GeoStats.jl/blob/master/LICENSE)
 [![Gitter](https://img.shields.io/badge/chat-on%20gitter-bc0067.svg)](https://gitter.im/JuliaEarth/GeoStats.jl)
 
+Cite as [![DOI](https://zenodo.org/badge/33827844.svg)](https://zenodo.org/badge/latestdoi/33827844)
+
 *High-performance implementations of geostatistical algorithms for the Julia programming language.*
 
 ## Overview
