### 2018-04-25 73a497c
Update community section in docs
