+++ docs/src/about/author.md
@@ -1,10 +1,0 @@
-**Júlio Hoffimann Mendes**
-
-I am a Ph.D. candidate in the Department of Energy Resources
-Engineering at Stanford University. You can find more about
-my research on my [website](https://juliohm.github.io).
-Below are some ways that we can connect:
-
-- [ResearchGate](https://www.researchgate.net/profile/Julio_Mendes2)
-- [LinkedIn](https://www.linkedin.com/in/júlio-hoffimann-834936116)
-- [GitHub](https://github.com/juliohm)
