### 2018-04-25 b3ca59f
Update contributing section in docs
Former-commit-id: b2d930e818feee4a26cd67f8870661b8587bba61
