### 2018-04-25 #692 @arfon
@juliohm - I had to fiddle around a little with the YAML header of your paper to get it to compile properly for production. In particular, I had to edit the email you had listed and turn it into a 'corresponding author' footnote:

```
---
title: 'GeoStats.jl -- High-performance geostatistics in Julia'
tags:
  - Geostatistics
  - Inverse problems
  - Random processes
  - Spatial statistics
  - Variography
  - Kriging
  - Estimation
  - Simulation
  - Julia
authors:
  - name: "Júlio Hoffimann^[Corresponding author: juliohm@stanford.edu]"
    orcid: 0000-0003-2789-297X
    affiliation: 1
affiliations:
  - name: Stanford University
    index: 1
date: 20 April 2018
bibliography: paper.bib
---

```
