### 2018-04-25 #692 @arfon
> Thank you all for the incredible process! Please let me know how I can become a reviewer for Julia software.

Great! Please fill in your details/interests here: http://joss.theoj.org/reviewer-signup.html
