### 2018-04-25 #692 @arfon
> The author name is listed incorrectly, and the cross references are not assigned to my profile. Is there any way to maintain the email but remove this side effect produced by LaTeX syntax?

There is not sorry. I think this is a limitation of our current template and document processing. The short-term fix here is to simply remove the footnote/your email from the submission.
