### 2018-04-25 #692 @juliohm
Thank you @arfon I will consider the change for future releases of the project.
